Compute RH from pressure, temperature and specific humidity
Add the diagnostic computation of relative humidity from pressure level,
temperature, and specific humidity.

The calculations to derive saturation vapor mixing ratio with respect to
liquid water are based on the RSLF function in the Thompson microphysics
scheme found in WRF. A simplified version of relative humidity
calculation is kept in the comments.