Enable only a subset of variables to be processed using the -v/--variable option
This commit adds the ability to process only a subset of the standard set of
variables that are handled by the era5_to_int.py script. The new -v/--variable
option may be given a comma-separated list, without spaces, of WPS variable
names to process. Only those specified variables will be written to the output
intermediate files.

Note that if a diagnostic variable (e.g, GHT) is needed in the output
intermediate files, all of the fields required to compute that diagnostic field
(e.g., GEOPT) must be specified in the list supplied to the -v/--variable
option.